feat(app-defaults): rename app-react plugin to app-defaults
Consolidate the app-react plugin into app-defaults, moving all drawer
components, extensions, hooks, and utilities. Updates all consumer
imports in the app package accordingly.
