+++ workspaces/app-defaults/packages/app/package.json
@@ -45,8 +45,8 @@
     "@material-ui/icons": "^4.9.1",
     "@mui/material": "^5.18.0",
     "@red-hat-developer-hub/backstage-plugin-app-auth": "workspace:^",
+    "@red-hat-developer-hub/backstage-plugin-app-defaults": "workspace:^",
     "@red-hat-developer-hub/backstage-plugin-app-integrations": "workspace:^",
-    "@red-hat-developer-hub/backstage-plugin-app-react": "workspace:^",
     "@red-hat-developer-hub/backstage-plugin-global-header": "^1.21.0",
     "material-icons": "^1.13.14",
     "react": "^18.0.2",
