+++ workspaces/app-defaults/plugins/app-auth/knip-report.md
@@ -1,0 +1,9 @@
+# Knip report
+
+## Unused devDependencies (2)
+
+| Name                   | Location          | Severity |
+| :--------------------- | :---------------- | :------- |
+| @testing-library/react | package.json:64:6 | error    |
+| @backstage/test-utils  | package.json:62:6 | error    |
+
