+++ workspaces/app-defaults/plugins/app-defaults/CHANGELOG.md
@@ -1,3 +1,5 @@
+# @red-hat-developer-hub/backstage-plugin-app-defaults
+
 # @red-hat-developer-hub/backstage-plugin-app-react
 
 ## 0.0.5
