+++ workspaces/app-defaults/plugins/app-defaults/knip-report.md
@@ -1,0 +1,8 @@
+# Knip report
+
+## Unused devDependencies (1)
+
+| Name                  | Location          | Severity |
+| :-------------------- | :---------------- | :------- |
+| @backstage/test-utils | package.json:45:6 | error    |
+
