+++ workspaces/app-defaults/plugins/app-defaults/report.api.md
@@ -1,0 +1,108 @@
+## API Report File for "@red-hat-developer-hub/backstage-plugin-app-defaults"
+
+> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
+
+```ts
+import { AppDrawerContent as AppDrawerContent_2 } from '@red-hat-developer-hub/backstage-plugin-app-react';
+import { ConfigurableExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
+import { ExtensionInput } from '@backstage/frontend-plugin-api';
+import { FrontendModule } from '@backstage/frontend-plugin-api';
+import { OverridableExtensionDefinition } from '@backstage/frontend-plugin-api';
+import { ReactNode } from 'react';
+
+// @public
+const appDefaults: FrontendModule;
+export default appDefaults;
+
+// @public
+export interface AppDrawerApi {
+  activeDrawerId: string | null;
+  closeDrawer(id: string): void;
+  getWidth(id: string): number;
+  isOpen(id: string): boolean;
+  openDrawer(id: string): void;
+  setWidth(id: string, width: number): void;
+  toggleDrawer(id: string): void;
+}
+
+// @public
+export interface AppDrawerContent {
+  defaultWidth?: number;
+  element: React.ReactElement;
+  id: string;
+  priority?: number;
+  resizable?: boolean;
+}
+
+// @public
+export const appDrawerExtension: OverridableExtensionDefinition<{
+  config: {};
+  configInput: {};
+  output: ExtensionDataRef<
+    (props: { children: ReactNode }) => JSX.Element | null,
+    'app.root.wrapper',
+    {}
+  >;
+  inputs: {
+    drawers: ExtensionInput<
+      ConfigurableExtensionDataRef<
+        AppDrawerContent_2,
+        'app.drawer.content',
+        {}
+      >,
+      {
+        singleton: false;
+        optional: false;
+        internal: false;
+      }
+    >;
+  };
+  kind: 'app-root-wrapper';
+  name: 'drawer';
+  params: {
+    Component?: [error: 'Use the `component` parameter instead'];
+    component: (props: { children: ReactNode }) => JSX.Element | null;
+  };
+}>;
+
+// Warning: (ae-missing-release-tag) "appDrawerExtensions" is part of the package's API, but it is missing a release tag (@alpha, @beta, @public, or @internal)
+//
+// @public (undocumented)
+export const appDrawerExtensions: OverridableExtensionDefinition<{
+  config: {};
+  configInput: {};
+  output: ExtensionDataRef<
+    (props: { children: ReactNode }) => JSX.Element | null,
+    'app.root.wrapper',
+    {}
+  >;
+  inputs: {
+    drawers: ExtensionInput<
+      ConfigurableExtensionDataRef<
+        AppDrawerContent_2,
+        'app.drawer.content',
+        {}
+      >,
+      {
+        singleton: false;
+        optional: false;
+        internal: false;
+      }
+    >;
+  };
+  kind: 'app-root-wrapper';
+  name: 'drawer';
+  params: {
+    Component?: [error: 'Use the `component` parameter instead'];
+    component: (props: { children: ReactNode }) => JSX.Element | null;
+  };
+}>[];
+
+// @public
+export const useAppDrawer: () => AppDrawerApi;
+
+// Warnings were encountered during analysis:
+//
+// src/drawer/extensions.d.ts:32:22 - (ae-undocumented) Missing documentation for "appDrawerExtensions".
+```
