+++ workspaces/app-defaults/plugins/app-defaults/src/drawer/components/ApplicationDrawer.tsx
