+++ workspaces/app-defaults/plugins/app-defaults/src/drawer/components/DrawerPanel.tsx
