+++ workspaces/app-defaults/plugins/app-integrations/knip-report.md
@@ -1,0 +1,9 @@
+# Knip report
+
+## Unused devDependencies (2)
+
+| Name                   | Location          | Severity |
+| :--------------------- | :---------------- | :------- |
+| @testing-library/react | package.json:60:6 | error    |
+| @backstage/test-utils  | package.json:58:6 | error    |
+
