--- workspaces/app-defaults/plugins/app-react/knip-report.md
@@ -1,14 +1,0 @@
-# Knip report
-
-## Unused dependencies (1)
-
-| Name                       | Location          | Severity |
-| :------------------------- | :---------------- | :------- |
-| @backstage/core-plugin-api | package.json:49:6 | error    |
-
-## Unused devDependencies (1)
-
-| Name                  | Location          | Severity |
-| :-------------------- | :---------------- | :------- |
-| @backstage/test-utils | package.json:59:6 | error    |
-
