--- workspaces/app-defaults/plugins/app-react/report-alpha.api.md
@@ -1,100 +1,0 @@
-## API Report File for "@red-hat-developer-hub/backstage-plugin-app-react"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-import { ConfigurableExtensionDataRef } from '@backstage/frontend-plugin-api';
-import { ExtensionBlueprint } from '@backstage/frontend-plugin-api';
-import { ExtensionDataRef } from '@backstage/frontend-plugin-api';
-import { FrontendModule } from '@backstage/frontend-plugin-api';
-import { JSX as JSX_2 } from 'react/jsx-runtime';
-
-// @public
-export interface AppDrawerApi {
-  activeDrawerId: string | null;
-  closeDrawer(id: string): void;
-  getWidth(id: string): number;
-  isOpen(id: string): boolean;
-  openDrawer(id: string): void;
-  setWidth(id: string, width: number): void;
-  toggleDrawer(id: string): void;
-}
-
-// @public
-export interface AppDrawerContent {
-  defaultWidth?: number;
-  element: React.ReactElement;
-  id: string;
-  priority?: number;
-  resizable?: boolean;
-}
-
-// @alpha
-export const AppDrawerContentBlueprint: ExtensionBlueprint<{
-  kind: 'app-drawer-content';
-  params: {
-    id: string;
-    element: React.ReactElement;
-    resizable?: boolean;
-    defaultWidth?: number;
-    priority?: number;
-  };
-  output: ExtensionDataRef<AppDrawerContent, 'app.drawer.content', {}>;
-  inputs: {};
-  config: {
-    defaultWidth: number | undefined;
-    resizable: boolean | undefined;
-    priority: number | undefined;
-  };
-  configInput: {
-    defaultWidth?: number | undefined;
-    resizable?: boolean | undefined;
-    priority?: number | undefined;
-  };
-  dataRefs: {
-    content: ConfigurableExtensionDataRef<
-      AppDrawerContent,
-      'app.drawer.content',
-      {}
-    >;
-  };
-}>;
-
-// @alpha
-export const appDrawerContentDataRef: ConfigurableExtensionDataRef<
-  AppDrawerContent,
-  'app.drawer.content',
-  {}
->;
-
-// @alpha
-export const appDrawerModule: FrontendModule;
-
-// @public
-export const ApplicationDrawer: (
-  input: ApplicationDrawerProps,
-) => JSX_2.Element;
-
-// @public
-export interface ApplicationDrawerProps {
-  children: React.ReactNode;
-  contents: AppDrawerContent[];
-}
-
-// @public
-export const DrawerPanel: (props: DrawerPanelProps) => JSX_2.Element;
-
-// @public
-export interface DrawerPanelProps {
-  children: React.ReactNode;
-  drawerWidth?: number;
-  isDrawerOpen: boolean;
-  isResizable?: boolean;
-  maxWidth?: number;
-  minWidth?: number;
-  onWidthChange?: (width: number) => void;
-}
-
-// @public
-export const useAppDrawer: () => AppDrawerApi;
-```
