--- workspaces/app-defaults/plugins/app-react/report.api.md
@@ -1,55 +1,0 @@
-## API Report File for "@red-hat-developer-hub/backstage-plugin-app-react"
-
-> Do not edit this file. It is a report generated by [API Extractor](https://api-extractor.com/).
-
-```ts
-import { JSX as JSX_2 } from 'react/jsx-runtime';
-
-// @public
-export interface AppDrawerApi {
-  activeDrawerId: string | null;
-  closeDrawer(id: string): void;
-  getWidth(id: string): number;
-  isOpen(id: string): boolean;
-  openDrawer(id: string): void;
-  setWidth(id: string, width: number): void;
-  toggleDrawer(id: string): void;
-}
-
-// @public
-export interface AppDrawerContent {
-  defaultWidth?: number;
-  element: React.ReactElement;
-  id: string;
-  priority?: number;
-  resizable?: boolean;
-}
-
-// @public
-export const ApplicationDrawer: (
-  input: ApplicationDrawerProps,
-) => JSX_2.Element;
-
-// @public
-export interface ApplicationDrawerProps {
-  children: React.ReactNode;
-  contents: AppDrawerContent[];
-}
-
-// @public
-export const DrawerPanel: (props: DrawerPanelProps) => JSX_2.Element;
-
-// @public
-export interface DrawerPanelProps {
-  children: React.ReactNode;
-  drawerWidth?: number;
-  isDrawerOpen: boolean;
-  isResizable?: boolean;
-  maxWidth?: number;
-  minWidth?: number;
-  onWidthChange?: (width: number) => void;
-}
-
-// @public
-export const useAppDrawer: () => AppDrawerApi;
-```
