--- workspaces/app-defaults/plugins/app-react/src/drawer/index.ts
@@ -1,27 +1,0 @@
-/*
- * Copyright Red Hat, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-export { useAppDrawer } from './hooks/useAppDrawer';
-export { ApplicationDrawer } from './components/ApplicationDrawer';
-export { DrawerPanel } from './components/DrawerPanel';
-
-export { appDrawerContentDataRef } from './extensions/appDrawerContentDataRef';
-export { AppDrawerContentBlueprint } from './extensions/AppDrawerContentBlueprint';
-export { appDrawerModule } from './extensions/appDrawerModule';
-
-export type { ApplicationDrawerProps } from './components/ApplicationDrawer';
-export type { DrawerPanelProps } from './components/DrawerPanel';
-export type { AppDrawerContent, AppDrawerApi } from './types';
